+++ src/commands/set.js
@@ -7,6 +7,10 @@ command.execute = async function (message, args) {
   const guildId = message.guild.id;
   const [soundId, soundUrl] = args;
 
+  if (!soundId || !soundUrl) {
+    throw new Error("!soundId || !soundUrl");
+  };
+
   await Sound.destroy({
     where: {
       guildId,
